Make p1_weather easier to understand.
Make p1_weather itself print out the values of calling the functions on
the default data when you run it.

Add tests.  For now, they only try the function on the default data
and check that they get the correct result.

Also add a README describing how to run things.